moves logic for re-opening modal on warning/error
