+++ src/GlobalConnectWalletModal.tsx
@@ -2,8 +2,15 @@ import {lazy, Suspense, useState} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import {useEffect} from 'react';
 
-import {connectModalClose} from './store/actions';
+import {connectModalClose, connectModalOpen} from './store/actions';
 import {StoreState} from './store/types';
+import {AsyncStatus} from './util/types';
+import {useLocation} from 'react-router-dom';
+import {
+  useIsDefaultChain,
+  useMaybeContractWallet,
+  useWeb3Modal,
+} from './components/web3/hooks';
 
 /**
  * Lazy load ConnectWalletModal only when open
@@ -22,12 +29,6 @@ export default function GlobalConnectWalletModal() {
   );
 
   /**
-   * Their hooks
-   */
-
-  const dispatch = useDispatch();
-
-  /**
    * State
    */
 
@@ -41,12 +42,60 @@ export default function GlobalConnectWalletModal() {
   });
 
   /**
+   * Our hooks
+   */
+
+  const {account, initialCachedConnectorCheckStatus, web3Instance} =
+    useWeb3Modal();
+
+  const {isDefaultChain} = useIsDefaultChain();
+
+  const maybeContractWallet = useMaybeContractWallet(
+    account,
+    web3Instance?.currentProvider
+  );
+
+  /**
+   * Their hooks
+   */
+
+  const dispatch = useDispatch();
+  const {pathname} = useLocation();
+
+  /**
    * Effects
    */
 
   useEffect(() => {
     setModalProps((prevState) => ({...prevState, isOpen: isConnectModalOpen}));
   }, [isConnectModalOpen]);
+
+  /**
+   * Effects
+   */
+
+  /**
+   * If the `web3Modal` is ready, and/or the `pathname` changes,
+   * then open or close the modal based on the current chain.
+   *
+   * When open, the user will be alerted to change chains.
+   */
+  useEffect(() => {
+    if (initialCachedConnectorCheckStatus === AsyncStatus.FULFILLED) {
+      const shouldShowModalAgain: boolean =
+        !isDefaultChain || maybeContractWallet ? true : false;
+
+      shouldShowModalAgain
+        ? dispatch(connectModalOpen())
+        : dispatch(connectModalClose());
+    }
+  }, [
+    dispatch,
+    initialCachedConnectorCheckStatus,
+    isDefaultChain,
+    maybeContractWallet,
+    pathname,
+  ]);
 
   /**
    * Render
