+++ src/components/web3/Web3ModalButton.tsx
@@ -1,12 +1,9 @@
 import {isMobile} from 'react-device-detect';
 import {useDispatch} from 'react-redux';
-import {useEffect} from 'react';
-import {useLocation} from 'react-router';
 
-import {AsyncStatus} from '../../util/types';
-import {connectModalClose, connectModalOpen} from '../../store/actions';
+import {connectModalOpen} from '../../store/actions';
 import {truncateEthAddress} from '../../util/helpers';
-import {useIsDefaultChain, useMaybeContractWallet} from './hooks';
+import {useIsDefaultChain} from './hooks';
 import {useWeb3Modal} from './hooks';
 import {WalletIcon} from './';
 
@@ -30,60 +27,21 @@ export default function ConnectWalletButton({
    * Our hooks
    */
 
-  const {
-    account,
-    connected,
-    initialCachedConnectorCheckStatus,
-    web3Modal,
-    web3Instance,
-  } = useWeb3Modal();
+  const {account, connected, web3Modal} = useWeb3Modal();
 
   const {isDefaultChain} = useIsDefaultChain();
-
-  const maybeContractWallet = useMaybeContractWallet(
-    account,
-    web3Instance?.currentProvider
-  );
 
   /**
    * Their hooks
    */
 
   const dispatch = useDispatch();
-  const {pathname} = useLocation();
 
   /**
    * Variables
    */
 
   const isWrongNetwork: boolean = isDefaultChain === false;
-
-  /**
-   * Effects
-   */
-
-  /**
-   * If the `web3Modal` is ready, and/or the `pathname` changes,
-   * then open or close the modal based on the current chain.
-   *
-   * When open, the user will be alerted to change chains.
-   */
-  useEffect(() => {
-    if (initialCachedConnectorCheckStatus === AsyncStatus.FULFILLED) {
-      const shouldShowModalAgain: boolean =
-        !isDefaultChain || maybeContractWallet ? true : false;
-
-      shouldShowModalAgain
-        ? dispatch(connectModalOpen())
-        : dispatch(connectModalClose());
-    }
-  }, [
-    dispatch,
-    initialCachedConnectorCheckStatus,
-    isDefaultChain,
-    maybeContractWallet,
-    pathname,
-  ]);
 
   /**
    * Functions
